Move db preperation from noxfile to conftest file(s)
